Visuals rearrange (#320)
* Spring force graph layout and example data.

* Marks-nesting of nodes.

* Mode reprsented as dounut chart showing attribution sources

* Link weight attribution WIP

* WIP: First draft rainbow edges

Actual offset/overlap varies by angle, but it exists.

* Layering and size-relationships improved.

* Node labels

* Multigraph visual added to the the plots module

* Color domain includes links.  Color legend added.

* Braking apart plots for more clear grouping of functions.

* Spring-graph: Label on top of nodes

* Mouse manipulation on multi-graph.

* Balanced stacking of edges.

* Parallel line equation
PROBLEM: alternate lines collapse around the NW/SE diagonal

* Multigraph working in all orientations (BUT slot is wrong...)
Mulitgraph working at all angular orientations

* Visual schema facet-check handes post-facet transformation.  Added legend to spring-graph

* Initial capability to save PNG.

* ipy_display produces a png by default.

Can do PNG, SVG or interactive.

* Highlight legend on chart-click

* Select exemplar trajectories by characteristics.
