+++ setup.cfg
@@ -25,7 +25,9 @@ install_requires =
     dask
     requests
     sympytorch
-        
+    vl-convert-python
+    statsmodels
+
 zip_safe = false
 include_package_data = true
 python_requires = >=3.9
