+++ src/pyciemss/visuals/schemas/calibrate_chart.vg.json
